Several changes and new functionality
ADD: Function to check if a Zombie entity is in game.
Needed to avoid the zombies getting in fire during sun time, but only
in the arenas of the game.

MODIFY: Ability to drop items in worlds that aren't arenas
In the original code it made imposible to drop objects in any world...

MODIFY: Ability to pick up blocks in worlds that aren't arenas
In the original code it won't work...

MODIFY: Translated ReadConfiguration file to English