Bug fixes, translations and debug mode
BUG: If config file doesn't exist, there can be running conditions
There can be problems loading configurations because the creation of
a new file occurs after trying to load it.

REMOVE: ForceRespawn dependency
It wasn't working, so the best option now is removing the dependency

ADD: Debug mode
In order to avoid filling the server console with SPAM every 5 seconds,
there is a new debug option in config.yml to enable detailed debugging.

MODIFY: Traducted to Engish and best programming style
GemShop and SelectKit were improved.

TODO: Fix ReSpawn and Leave events to complete the game functionality