Added spectator view and leave protection
ADD: Spectator view
Now if you died, you become a spectator and con flight, but other
users and the mobs can't see you.
When the next wave arrives, you come to life and continue gaming.

ADD: Leave protection
If you leave the game while you are playing, first you are teleported
to the Arena Lobby, so when you enter again you won't be locked in an
arena world.